Create twins.cpp
Take an input str, a string. A “twin” is defined as two instances of a char separated by a char.
E.g. "AxA" the A's make a “twin”. “twins” can overlap, so "AxAxA" contains 3 “twins” - 2 for A and 1 for x.
Write a function which recursively counts number of “twins” in a string. Print the value returned.